separate tracking db
